refactor(auth): Adapt AuthModule to use SecurityModule for password services
